requirements.txt
azure-functions==1.18.0
azure-identity==1.15.0
azure-storage-blob==12.19.0
azure-ai-formrecognizer==3.3.2
azure-search-documents==11.4.0
azure-cosmos==4.5.1
openai==1.12.0
azure-monitor-opentelemetry==1.2.0
opencensus-ext-azure==1.1.13
applicationinsights==0.11.10
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
jinja2==3.1.3
pillow==10.2.0
numpy==1.26.4
pandas==2.2.0
tiktoken==0.6.0
tenacity==8.2.3
pyodbc==5.0.1
sqlalchemy==2.0.27
alembic==1.13.1
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0
faker==22.6.0
deepdiff==6.7.1
ragas==0.1.5
langchain==0.1.7
langchain-openai==0.0.5
